requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
mlflow<3.0.0
PyYAML
dagshub
fastapi
uvicorn
python-multipart
setuptools

##-e .